Added Score Penalties to the Prefabs
Prefabs now have a public variable for how much they will decrease the player's score when destroyed by fire. I also added a custom datatype enum for prefabs that allows us to specify their object type in the editor "tree/building" etc. This datatype is used to increase the number of destroyed trees/buildings in the Game Manager which is a stat we can show at the end.

Also added a toggle for the minimap in phase 2 with the 'M' key and changed the debug keys for loading scenes in the phase controllers to the 1,2 and 3 numbers keys.