Added Debug Score Text to Phase 3
Added some super debug temporary score text to phase 3 so we can test the gameplay of phase 2 and check the difficulty balance. Also fixed a bug where the prefabs were constantly updating the score every frame.

Increased the number of initial fires made by the weather system too and added a firestarter function to it so that it makes new fires every 30 seconds.